+++ APSITests/src/channel_tests.cpp
@@ -389,21 +389,21 @@ namespace APSITests
         ASSERT_EQ(true, get_params_response.psiconf_params.use_oprf);
         ASSERT_EQ(true, get_params_response.psiconf_params.use_labels);
         ASSERT_EQ(false, get_params_response.psiconf_params.use_fast_membership);
-        ASSERT_EQ((unsigned)60, get_params_response.psiconf_params.item_bit_count);
-        ASSERT_EQ((unsigned)120, get_params_response.psiconf_params.item_bit_length_used_after_oprf);
-        ASSERT_EQ((unsigned)40, get_params_response.psiconf_params.num_chunks);
-        ASSERT_EQ((unsigned)50, get_params_response.psiconf_params.sender_bin_size);
-        ASSERT_EQ((unsigned)10, get_params_response.table_params.log_table_size);
-        ASSERT_EQ((unsigned)1, get_params_response.table_params.window_size);
-        ASSERT_EQ((unsigned)2, get_params_response.table_params.split_count);
-        ASSERT_EQ((unsigned)10, get_params_response.table_params.split_size);
-        ASSERT_EQ((unsigned)40, get_params_response.table_params.binning_sec_level);
-        ASSERT_EQ((unsigned)3, get_params_response.cuckoo_params.hash_func_count);
-        ASSERT_EQ((unsigned)2, get_params_response.cuckoo_params.hash_func_seed);
-        ASSERT_EQ((unsigned)1, get_params_response.cuckoo_params.max_probe);
+        ASSERT_EQ((u32)60, get_params_response.psiconf_params.item_bit_count);
+        ASSERT_EQ((u32)120, get_params_response.psiconf_params.item_bit_length_used_after_oprf);
+        ASSERT_EQ((u32)40, get_params_response.psiconf_params.num_chunks);
+        ASSERT_EQ((u32)50, get_params_response.psiconf_params.sender_bin_size);
+        ASSERT_EQ((u32)10, get_params_response.table_params.log_table_size);
+        ASSERT_EQ((u32)1, get_params_response.table_params.window_size);
+        ASSERT_EQ((u32)2, get_params_response.table_params.split_count);
+        ASSERT_EQ((u32)10, get_params_response.table_params.split_size);
+        ASSERT_EQ((u32)40, get_params_response.table_params.binning_sec_level);
+        ASSERT_EQ((u32)3, get_params_response.cuckoo_params.hash_func_count);
+        ASSERT_EQ((u32)2, get_params_response.cuckoo_params.hash_func_seed);
+        ASSERT_EQ((u32)1, get_params_response.cuckoo_params.max_probe);
         ASSERT_EQ((u64)678910, get_params_response.exfield_params.characteristic);
-        ASSERT_EQ((unsigned)8, get_params_response.exfield_params.degree);
-        ASSERT_EQ((unsigned)25, get_params_response.seal_params.max_supported_degree);
+        ASSERT_EQ((u32)8, get_params_response.exfield_params.degree);
+        ASSERT_EQ((u32)25, get_params_response.seal_params.max_supported_degree);
         ASSERT_EQ((u64)5119, get_params_response.seal_params.encryption_params.plain_modulus().value());
         ASSERT_EQ((size_t)4096, get_params_response.seal_params.encryption_params.poly_modulus_degree());
         ASSERT_EQ((size_t)3, get_params_response.seal_params.encryption_params.coeff_modulus().size());
@@ -418,16 +418,16 @@ namespace APSITests
         ASSERT_EQ((u64)54321, get_params_response2.psiconf_params.sender_size);
         ASSERT_EQ(false, get_params_response2.psiconf_params.use_oprf);
         ASSERT_EQ(false, get_params_response2.psiconf_params.use_labels);
-        ASSERT_EQ((unsigned)80, get_params_response2.psiconf_params.item_bit_count);
-        ASSERT_EQ((unsigned)10, get_params_response2.table_params.log_table_size);
-        ASSERT_EQ((unsigned)1, get_params_response2.table_params.window_size);
-        ASSERT_EQ((unsigned)2, get_params_response2.table_params.split_count);
-        ASSERT_EQ((unsigned)40, get_params_response2.table_params.binning_sec_level);
-        ASSERT_EQ((unsigned)3, get_params_response2.cuckoo_params.hash_func_count);
-        ASSERT_EQ((unsigned)2, get_params_response2.cuckoo_params.hash_func_seed);
-        ASSERT_EQ((unsigned)1, get_params_response2.cuckoo_params.max_probe);
+        ASSERT_EQ((u32)80, get_params_response2.psiconf_params.item_bit_count);
+        ASSERT_EQ((u32)10, get_params_response2.table_params.log_table_size);
+        ASSERT_EQ((u32)1, get_params_response2.table_params.window_size);
+        ASSERT_EQ((u32)2, get_params_response2.table_params.split_count);
+        ASSERT_EQ((u32)40, get_params_response2.table_params.binning_sec_level);
+        ASSERT_EQ((u32)3, get_params_response2.cuckoo_params.hash_func_count);
+        ASSERT_EQ((u32)2, get_params_response2.cuckoo_params.hash_func_seed);
+        ASSERT_EQ((u32)1, get_params_response2.cuckoo_params.max_probe);
         ASSERT_EQ((u64)678910, get_params_response2.exfield_params.characteristic);
-        ASSERT_EQ((unsigned)8, get_params_response2.exfield_params.degree);
+        ASSERT_EQ((u32)8, get_params_response2.exfield_params.degree);
         ASSERT_EQ((u64)5119, get_params_response2.seal_params.encryption_params.plain_modulus().value());
         ASSERT_EQ((size_t)4096, get_params_response2.seal_params.encryption_params.poly_modulus_degree());
         ASSERT_EQ((size_t)3, get_params_response2.seal_params.encryption_params.coeff_modulus().size());
@@ -573,9 +573,9 @@ namespace APSITests
             });
 
         vector<thread> clients(5);
-        for (unsigned i = 0; i < clients.size(); i++)
+        for (size_t i = 0; i < clients.size(); i++)
         {
-            clients[i] = thread([this](unsigned idx)
+            clients[i] = thread([this](size_t idx)
                 {
                     ReceiverChannel recv;
 
@@ -584,7 +584,7 @@ namespace APSITests
                     u8 a = static_cast<u8>(idx) * 2;
                     u8 b = a + 1;
 
-                    for (unsigned i = 0; i < 5; i++)
+                    for (u32 i = 0; i < 5; i++)
                     {
                         vector<u8> buffer(2);
                         buffer[0] = a;
@@ -602,7 +602,7 @@ namespace APSITests
                 }, i);
         }
 
-        for (unsigned i = 0; i < clients.size(); i++)
+        for (size_t i = 0; i < clients.size(); i++)
         {
             clients[i].join();
         }
